update package names
