+++ src/main/java/site/stellaburgers/order/OrderCreate.java
@@ -1,4 +1,4 @@
-package site.stellaburgers.Order;
+package site.stellaburgers.order;
 
 
 import java.util.List;
