+++ src/test/java/site/stellaburgers/userTest/UserCreateTest.java
@@ -1,4 +1,4 @@
-package site.stellaburgers.UserTest;
+package site.stellaburgers.userTest;
 
 
 import io.qameta.allure.Description;
@@ -7,9 +7,9 @@ import io.restassured.response.ValidatableResponse;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import site.stellaburgers.User.UserCreate;
-import site.stellaburgers.User.UserLogin;
-import site.stellaburgers.User.UserStep;
+import site.stellaburgers.user.UserCreate;
+import site.stellaburgers.user.UserLogin;
+import site.stellaburgers.user.UserStep;
 
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.Matchers.equalTo;
